+++ packages/ember-data/lib/transforms/base.js
@@ -1,0 +1,7 @@
+DS.Transform = Ember.Object.extend({
+
+  serialize: Ember.required(),
+  
+  deserialize: Ember.required()
+
+});
