+++ packages/ember-data/lib/transforms/index.js
@@ -1,0 +1,5 @@
+require('ember-data/transforms/base');
+require('ember-data/transforms/boolean');
+require('ember-data/transforms/date');
+require('ember-data/transforms/number');
+require('ember-data/transforms/string');
