+++ packages/ember-data/lib/transforms/string.js
@@ -1,0 +1,13 @@
+var none = Ember.isNone, empty = Ember.isEmpty;
+
+DS.StringTransform = DS.Transform.extend({
+
+  deserialize: function(serialized) {
+    return none(serialized) ? null : String(serialized);
+  },
+
+  serialize: function(deserialized) {
+    return none(deserialized) ? null : String(deserialized);
+  }
+
+});
