+++ tests/ember_configuration.js
@@ -173,6 +173,20 @@
       updateRecordArraysLater: syncForTest()
     });
 
+    var transforms = {
+      'boolean': DS.BooleanTransform.create(),
+      'date': DS.DateTransform.create(),
+      'number': DS.NumberTransform.create(),
+      'string': DS.StringTransform.create()
+    };
+
+    // Prevent all tests involving serialization to require a container
+    DS.JSONSerializer.reopen({
+      transformFor: function(attributeType) {
+        return this._super(attributeType) || transforms[attributeType];
+      }
+    });
+
     Ember.RSVP.Promise.prototype.then = syncForTest(Ember.RSVP.Promise.prototype.then);
   });
